waay

+++ 0x15-file_io/3-cp.c
@@ -22,18 +22,18 @@ int main(int ac, char **av)
 	if (ac != 3)
 		fprintf(stderr, "Usage: cp file_from file_to\n"), exit(97);
 	ofile_from = open(av[1], O_RDONLY); /*open file from "the first file" */
-	if (ofile_from == -1)
+	if ((ssize_t)ofile_from == -1)
 		fprintf(stderr, "Error: Can't read from file %s\n", av[1]), exit(98);
 
 	ofile_to = open(av[2], O_CREAT | O_RDWR | O_TRUNC, PER);
-	if (ofile_to == -1)
+	if ((ssize_t)ofile_to == -1)
 		fprintf(stderr, "Error: Can't write to %s\n", av[2]), exit(99);
 
 	while ((rbuff =  read(ofile_from, buff, 1024)) > 0)
 		if (write(ofile_to, buff, rbuff) != rbuff)
 			fprintf(stderr, "Error: Can't read from file %s\n", av[1]), exit(98);
 
-	if (rbuff == -1)
+	if ((ssize_t)rbuff == -1)
 		fprintf(stderr, "Error: Can't read from file %s\n", av[1]), exit(98);
 
 	ofile_from = close(ofile_from);
